Move tests to a designated file
