+++ src/geo/projection/globe_transform.test.ts
@@ -9,7 +9,6 @@ import {angularCoordinatesRadiansToVector, mercatorCoordinatesToAngularCoordinat
 import {expectToBeCloseToArray, getGlobeProjectionMock, sleep} from '../../util/test/util';
 import {MercatorCoordinate} from '../mercator_coordinate';
 import {tileCoordinatesToLocation} from './mercator_utils';
-import {coveringTiles} from './covering_tiles';
 
 function testPlaneAgainstLngLat(lngDegrees: number, latDegrees: number, plane: Array<number>) {
     const lat = latDegrees / 180.0 * Math.PI;
@@ -609,166 +608,6 @@ describe('GlobeTransform', () => {
 
         test('barely hidden', () => {
             expect(transform.isLocationOccluded(new LngLat(84.50, 0))).toBe(true);
-        });
-    });
-
-    describe('coveringTiles', () => {
-        test('zoomed out', () => {
-            const transform = new GlobeTransform(globeProjectionMock);
-            transform.resize(128, 128);
-            transform.setCenter(new LngLat(0.0, 0.0));
-            transform.setZoom(-1);
-
-            const tiles = coveringTiles(transform, {
-                tileSize: 512,
-            });
-
-            expect(tiles).toEqual([
-                new OverscaledTileID(0, 0, 0, 0, 0)
-            ]);
-        });
-
-        test('zoomed in', () => {
-            const transform = new GlobeTransform(globeProjectionMock);
-            transform.resize(128, 128);
-            transform.setCenter(new LngLat(-0.02, 0.01));
-            transform.setZoom(3);
-
-            const tiles = coveringTiles(transform, {
-                tileSize: 512,
-            });
-
-            expect(tiles).toEqual([
-                new OverscaledTileID(3, 0, 3, 3, 3),
-                new OverscaledTileID(3, 0, 3, 3, 4),
-                new OverscaledTileID(3, 0, 3, 4, 3),
-                new OverscaledTileID(3, 0, 3, 4, 4),
-            ]);
-        });
-
-        test('zoomed in 512x512', () => {
-            const transform = new GlobeTransform(globeProjectionMock);
-            transform.resize(512, 512);
-            transform.setCenter(new LngLat(-0.02, 0.01));
-            transform.setZoom(3);
-
-            const tiles = coveringTiles(transform, {
-                tileSize: 512,
-            });
-
-            expect(tiles).toEqual([
-                new OverscaledTileID(3, 0, 3, 3, 3),
-                new OverscaledTileID(3, 0, 3, 3, 4),
-                new OverscaledTileID(3, 0, 3, 4, 3),
-                new OverscaledTileID(3, 0, 3, 4, 4),
-                new OverscaledTileID(3, 0, 3, 2, 3),
-                new OverscaledTileID(3, 0, 3, 2, 4),
-                new OverscaledTileID(3, 0, 3, 5, 3),
-                new OverscaledTileID(3, 0, 3, 5, 4),
-                new OverscaledTileID(3, 0, 3, 2, 2),
-                new OverscaledTileID(3, 0, 3, 2, 5),
-                new OverscaledTileID(3, 0, 3, 5, 2),
-                new OverscaledTileID(3, 0, 3, 5, 5),
-            ]);
-        });
-
-        test('pitched', () => {
-            const transform = new GlobeTransform(globeProjectionMock);
-            transform.resize(128, 128);
-            transform.setCenter(new LngLat(-0.002, 0.001));
-            transform.setZoom(8);
-            transform.setMaxPitch(80);
-            transform.setPitch(80);
-
-            const tiles = coveringTiles(transform, {
-                tileSize: 512,
-            });
-
-            expect(tiles).toEqual([
-                new OverscaledTileID(6, 0, 6, 32, 31),
-                new OverscaledTileID(6, 0, 6, 31, 31),
-                new OverscaledTileID(10, 0, 10, 511, 512),
-                new OverscaledTileID(10, 0, 10, 512, 512)
-            ]);
-        });
-
-        test('pitched+rotated', () => {
-            const transform = new GlobeTransform(globeProjectionMock);
-            transform.resize(128, 128);
-            transform.setCenter(new LngLat(-0.002, 0.001));
-            transform.setZoom(8);
-            transform.setMaxPitch(80);
-            transform.setPitch(80);
-            transform.setBearing(45);
-
-            const tiles = coveringTiles(transform, {
-                tileSize: 512,
-            });
-
-            expect(tiles).toEqual([
-                new OverscaledTileID(7, 0, 7, 64, 64),
-                new OverscaledTileID(7, 0, 7, 65, 63),
-                new OverscaledTileID(7, 0, 7, 64, 63),
-                new OverscaledTileID(7, 0, 7, 63, 63),
-                new OverscaledTileID(7, 0, 7, 64, 62),
-                new OverscaledTileID(10, 0, 10, 510, 512),
-                new OverscaledTileID(10, 0, 10, 511, 512),
-                new OverscaledTileID(10, 0, 10, 511, 513)
-            ]);
-        });
-
-        test('antimeridian1', () => {
-            const transform = new GlobeTransform(globeProjectionMock);
-            transform.resize(128, 128);
-            transform.setCenter(new LngLat(179.99, -0.001));
-            transform.setZoom(5);
-
-            const tiles = coveringTiles(transform, {
-                tileSize: 512,
-            });
-
-            expect(tiles).toEqual([
-                new OverscaledTileID(5, 0, 5, 31, 16),
-                new OverscaledTileID(5, 0, 5, 31, 15),
-                new OverscaledTileID(5, 1, 5, 0, 16),
-                new OverscaledTileID(5, 1, 5, 0, 15),
-            ]);
-        });
-
-        test('antimeridian2', () => {
-            const transform = new GlobeTransform(globeProjectionMock);
-            transform.resize(128, 128);
-            transform.setCenter(new LngLat(-179.99, 0.001));
-            transform.setZoom(5);
-
-            const tiles = coveringTiles(transform, {
-                tileSize: 512,
-            });
-
-            expect(tiles).toEqual([
-                new OverscaledTileID(5, 0, 5, 0, 15),
-                new OverscaledTileID(5, 0, 5, 0, 16),
-                new OverscaledTileID(5, -1, 5, 31, 15),
-                new OverscaledTileID(5, -1, 5, 31, 16),
-            ]);
-        });
-
-        test('zoom < 0', () => {
-            const transform = new GlobeTransform(globeProjectionMock);
-            transform.resize(128, 128);
-            transform.setCenter(new LngLat(0.0, 80.0));
-            transform.setZoom(-0.5);
-
-            const tiles = coveringTiles(transform, {
-                tileSize: 512,
-                minzoom: 0,
-                maxzoom: 0,
-                reparseOverscaled: true
-            });
-
-            expect(tiles).toEqual([
-                new OverscaledTileID(0, 0, 0, 0, 0)
-            ]);
         });
     });
